Refactor routes config
